V2.0 Big update.
*Filters
Tried to add a filter but ran into issue because of the multiple repositories. This commit changes how the components are listed.
Already one list to test search capabilities

*Categories
Changed componentlist to show only categories that will link to a list of a certain category like processors. So afterwards filtering will be easier to code.
Added queries to the repos to test search.

Added a customized icon pack to visual please this page.
Still need to add links and the pages it corresponds to.

*Components
Moved the detail pages to a folder to clean up the directory.

Changed some styling